refactor(editor): remove dead board prompt code
